renamed cli modules to have _cli suffix
