+++ xnat_ingest/cli/__init__.py
@@ -1,10 +1,10 @@
-from .assign import assign_cmd
-from .associate import associate_cmd
+from .assign_cli import assign_cmd
+from .associate_cli import associate_cmd
 from .base import cli
-from .check_upload import check_upload_cmd
-from .deidentify import deidentify_cmd
-from .group import group_cmd, group_orthanc_cmd
-from .upload import upload_cmd
+from .check_upload_cli import check_upload_cmd
+from .deidentify_cli import deidentify_cmd
+from .group_cli import group_cmd, group_orthanc_cmd
+from .upload_cli import upload_cmd
 
 __all__ = [
     "assign_cmd",
