+++ xnat_ingest/cli/assign_cli.py
