+++ xnat_ingest/cli/group_cli.py
