+++ xnat_ingest/cli/upload_cli.py
